+++ examples/basic_hyper/src/main.rs
@@ -1,3 +1,4 @@
+use bytes::Bytes;
 use gitlab_client::client::http::HttpGitlabClient;
 use gitlab_client::common::project::{ProjectRef, ProjectSlug};
 use gitlab_client::compact_str::CompactString;
@@ -7,7 +8,9 @@ use gitlab_client::query::get_tree_record_list::GetTreeRecordListQuery;
 use gitlab_client::tower_service::Service;
 use gitlab_client::url::Url;
 use gitlab_client::{GitlabAuth, UserAgent};
+use http_body_util::Full;
 use hyper_tls::HttpsConnector;
+use hyper_util::client::legacy::Client;
 
 #[tokio::main]
 async fn main() {
@@ -18,7 +21,8 @@ async fn main() {
   };
 
   let connector = HttpsConnector::new();
-  let client = hyper_util::client::legacy::Client::builder(hyper_util::rt::TokioExecutor::new()).build(connector);
+  let client: Client<HttpsConnector<_>, Full<Bytes>> =
+    Client::builder(hyper_util::rt::TokioExecutor::new()).build(connector);
   let mut client = HttpGitlabClient::new(client);
   let context = Context::new()
     .set_gitlab_url(GitlabUrl(Url::parse("https://gitlab.com/").unwrap()))