+++ src/query/mod.rs
@@ -5,4 +5,5 @@ pub mod get_project_list_page;
 pub mod get_project_package_list;
 pub mod get_project_release;
 pub mod get_project_release_list;
+pub mod get_project_release_list_page;
 pub mod get_tree_record_list;